+++ resources/serverless/values.yaml
@@ -167,7 +167,7 @@ containers:
       targetCPUUtilizationPercentage:
         value: "50"
       functionPublisherProxyAddress:
-        "http://eventing-publisher-proxy.kyma-system.svc.cluster.local/publish"
+        value: "http://eventing-publisher-proxy.kyma-system.svc.cluster.local/publish"
       functionRequeueDuration:
         value: 5m
       functionBuildExecutorArgs:
